pyproject.toml
[project]
name = "twitter-post-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.6.2",
    "nltk>=3.9.1",
    "python-dotenv>=1.1.1",
    "unstructured[md]>=0.18.11",
]
